feat(ui): add chapter and page icons to list items for better visual hierarchy & remove /books url from navbar
This commit introduces icons for chapters and pages in the list items to improve the visual hierarchy and user experience. The icons are styled with a consistent size and alignment. Additionally, the /books url is removed from navbar & default home page file was removed as it is no longer needed.